move data folder into lib
