+++ src/components/RouteIcon.svelte
@@ -1,13 +1,13 @@
 <script lang="ts">
-  import type { components } from '../data/bkk-openapi';
+  import type { components } from '../lib/data/bkk-openapi';
 
   export let routeData: components['schemas']['TransitRoute'] = {};
 </script>
 
-  <span
-    style:color={'#' + routeData?.style?.icon?.textColor}
-    style:background-color={'#' + routeData?.style?.color}
-    class="rounded p-1"
-  >
-    {routeData?.shortName}
-  </span>
+<span
+  style:color={'#' + routeData?.style?.icon?.textColor}
+  style:background-color={'#' + routeData?.style?.color}
+  class="rounded p-1"
+>
+  {routeData?.shortName}
+</span>
